requirements.txt
alembic==1.8.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.6.1
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-celery==2.1.4
apache-airflow-providers-common-sql==1.9.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-postgres==4.1.0
apache-airflow-providers-redis==2.0.4
apache-airflow-providers-sqlite==2.1.3
apispec==6.3.0
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.10.1
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.1.0
blinker==1.4
cachelib==0.7.0
cattrs==23.2.3
celery==5.2.7
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
comm==0.2.0
commonmark==0.9.1
ConfigUpdater==3.2
connexion==2.13.1
contourpy==1.2.0
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==1.0.0
fonttools==4.46.0
fqdn==1.5.1
google-re2==1.1
googleapis-common-protos==1.62.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
humanize==4.1.0
idna==3.3
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.0.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kombu==5.2.4
lazy-object-proxy==1.7.1
limits==3.7.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
overrides==7.4.0
packaging==21.3
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.12.0
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyspark==3.3.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==3.5.3
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.4.4
rich-argparse==1.4.0
rpds-py==0.13.2
Send2Trash==1.8.2
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.5
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.4
stack-data==0.6.3
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.8
widgetsnbextension==4.0.9
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
